Debug and minor changes
Had an issue where "Portfolio_Home" wasn't loading the CSS attached to the file. To remedy this, the file was then saved as a html file instead of a text file. Also a couple lines of code where edited to prepare the webpage to be organized with the flexbox method.